Replace print by the Python logging framework
Log messages with level >= ERROR give an event log
message.

Also: move scheduler to a separate file scheduler.py.
